requirements.txt
allure_python_commons==2.9.45
beautifulsoup4==4.11.1
pytest==7.1.2
requests==2.28.1
selenium==4.3.0
